Align navigation items in vertical center
Due to the chinese language icon on the right being slightly taller than the rest of the navigation items,
the alignment was slightly off (by two pixels or so). To the viewer with a high attention to such details,
this hurts the eyes. This makes the additional div a flex container as well and aligns the items in their
vertical center, regardless of their individual height.